GUI & User-error
Minor adjustments made to the GUI. It is now more compact, and the
ComboBoxes lengths have been adjusted so no text is cut-off for some of
the longer option names. Initial options in ComboBox2 removed- the
script handles these options automatically already (after moving
comboBox1's index initialization further down in the script)

TO-DO: TB 6A-8B, final testing, comments, a README, then it'll be time
to package this up and release it.

Selections now check if the first line of options passes the second line
chronologically and updates the second line accordingly. Debating
whether I want this to work both ways which would make it impossible to
choose an invalid range.
The Unit boxes now clear if the subject isn't VF.