SM
The old SM path returns to handle SM 6A and 6B, Sections for 7A-8A (as
well as the start of 8B) have also been added. This means every subject
worksheet is handled except for PH1 and the rest of SM 8B (both of which
are literally never used anyway so I'm not going to care)

Fixed the range for SV grades with only 9 lessons
Fixed a fatal bug where not having a Unit (aka, looking for any file
other than one in VF) would cause any file that relied on $FULLTYPE, to
fail.
Reverted LF's filename back to having the wildcard because it already
works and I probably wasn't saving any time without it anyways.

Given some time at work tomorrow, I'll probably try to do a "complete"
testing (opening literally every file), barring that, and some of the
formalities (comments, README, yadda yadda), this is done.