STAMS
Removed STAMS from the GUI as per manager's request from last week (the
code to find them is all there if they ever ask for it back). Also
removed some unused arrays from a dated WB 7A-8B implementation.
To-Do: Properly package this or codesign it as running it as an unsigned
script is a security vulnerability.